+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ThreadFacadeTest.java
@@ -27,7 +27,15 @@ public class ThreadFacadeTest {
         logger = new InterceptingLogger();
         List<Stackframe> frames = Collections.emptyList();
         stacktrace = new Stacktrace(frames);
-        thread = new Thread(1, "thread-2", ThreadType.ANDROID, false, stacktrace, logger);
+        thread = new Thread(
+                1,
+                "thread-2",
+                ThreadType.ANDROID,
+                false,
+                Thread.State.RUNNABLE,
+                stacktrace,
+                logger
+        );
     }
 
     @Test